[KEH-239] - Created new tab and initial function for demo
